Organise classes into appropriate Java packages
