+++ src/main/java/gilu/Ui.java
@@ -1,5 +1,7 @@
 package gilu;
 
+import gilu.task.Task;
+
 /**
  * Handles user interactions.
  */
