+++ src/main/java/gilu/task/Deadline.java
@@ -1,4 +1,4 @@
-package gilu;
+package gilu.task;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
