+++ src/main/java/gilu/task/Task.java
@@ -1,4 +1,4 @@
-package gilu;
+package gilu.task;
 
 /**
  * Represents a generic task with a description and completion status.
